Fix app layout
