+++ lib/primitives/Layout/index.tsx
@@ -1,2 +1,3 @@
 export * from "./AutoGrid"
+export * from "./Split"
 export * from "./Stack"
